SberCloud/task_2_trello/requirements.txt
pytest==3.6.2
selenium==3.141.0
webdriver-manager==3.5.0
allure-pytest==2.9.44